+++ docs/docs/.vitepress/utils/sidebar.ts
@@ -99,7 +99,7 @@ export const sidebar = {
         { text: 'Watermark 水印', link: '/components/watermark' },
         { text: 'Ripple 涟漪', link: '/components/ripple' },
         { text: 'Calendar 日历', link: '/components/calendar' },
-        { text: 'Waterfall 瀑布流', link: '/components/waterfall' }
+        // { text: 'Waterfall 瀑布流', link: '/components/waterfall' }
       ]
     }
   ]