+++ start/src/App.vue
@@ -1,14 +1,25 @@
 <template>
-  <f-button @click="openMessage">查看消息</f-button>
+  <f-checkbox-group v-model="option4" border size="large">
+    <f-checkbox label="选项一" />
+    <f-checkbox label="选项二" />
+    <f-checkbox label="选项三" />
+  </f-checkbox-group>
+
+  <f-checkbox-group v-model="option4" border>
+    <f-checkbox label="选项一" />
+    <f-checkbox label="选项二" />
+    <f-checkbox label="选项三" />
+  </f-checkbox-group>
+
+  <f-checkbox-group v-model="option4" border size="small">
+    <f-checkbox label="选项一" />
+    <f-checkbox label="选项二" />
+    <f-checkbox label="选项三" />
+  </f-checkbox-group>
 </template>
 
 <script lang="ts" setup>
-  import { FMessage } from '../../packages/fighting-design/index'
+  import { ref } from 'vue'
 
-  const openMessage = () => {
-    const info = FMessage({
-      message: '这是一条消息提示',
-      type: 'success'
-    })
-  }
+  const option4 = ref([])
 </script>